pyproject.toml
[project]
name = "ccmux"
version = "0.1.0"
description = "Telegram Bot for monitoring Claude Code sessions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
ccmux = "ccmux.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ccmux"]
